adds tests for calcP and modifies util
