+++ bet/calculateP/calculateP.py
@@ -64,7 +64,8 @@ def prob_emulated(samples, data, rho_D_M, d_distr_samples, lam_domain,
     :returns: (P, lambda_emulate, io_ptr, emulate_ptr, lam_vol)
 
     """
-    
+    if len(data.shape) == 1:
+        data = np.expand_dims(data, axis=1) 
     if lambda_emulate == None:
         lambda_emulate = samples
     if len(d_distr_samples.shape) == 1:
@@ -124,20 +125,26 @@ def prob(samples, data, rho_D_M, d_distr_samples, lam_domain, d_Tree=None):
         a pointer from emulated samples to samples (in parameter space)
 
     """
+    if len(data.shape) == 1:
+        data = np.expand_dims(data, axis=1) 
+
     # Calculate pointers and volumes
     (P, _, io_ptr, emulate_ptr) = prob_emulated(samples, data,
             rho_D_M, d_distr_samples, lam_domain, None, d_Tree)
     
     # Apply the standard MC approximation 
-    lam_vol = np.ones((samples.shape[0],))
+    lam_vol = np.ones((samples.shape[0],1))
     # Calculate Probabilities
     P = np.zeros((samples.shape[0],))
+    # TODO Fix issue of unecessary mpi call
+
     for i in range(rho_D_M.shape[0]):
         Itemp = np.equal(io_ptr, i)
         Itemp_sum = np.sum(lam_vol[Itemp])
         Itemp_sum = comm.allreduce(Itemp_sum, op=MPI.SUM)
         if Itemp_sum > 0:
             P[Itemp] = rho_D_M[i]*lam_vol[Itemp]/Itemp_sum 
+    lam_vol = (1.0/float(samples.shape[0]))*np.ones((samples.shape[0],1))
 
     return (P, lam_vol, io_ptr, emulate_ptr)
 
@@ -213,7 +220,7 @@ def prob_qhull(samples, data, rho_D_M, d_distr_samples,
     
     # Calculate the Voronoi diagram for samples. Calculate the volumes of 
     # the convex hulls of the corresponding Voronoi regions.
-    lam_vol = np.zeros((samples.shape[-1],))
+    lam_vol = np.zeros((samples.shape[-1],1))
     for i in range((samples.shape[0])):
         vornoi = spatial.Voronoi(lam_bound)
         lam_vol[i] = float(pyhull.qconvex('Qt FA', vornoi.vertices).split()[-1])
@@ -257,6 +264,8 @@ def prob_mc(samples, data, rho_D_M, d_distr_samples,
         a pointer from emulated samples to samples (in parameter space)
 
     """
+    if len(data.shape) == 1:
+        data = np.expand_dims(data, axis=1) 
     # Calculate pointers and volumes
     (P, lambda_emulate, io_ptr, emulate_ptr) = prob_emulated(samples,
             data, rho_D_M, d_distr_samples, lam_domain, lambda_emulate,
@@ -265,7 +274,7 @@ def prob_mc(samples, data, rho_D_M, d_distr_samples,
     # Apply the standard MC approximation to determine the number of emulated
     # samples per model run sample. This is for approximating 
     # \mu_Lambda(A_i \intersect b_j)
-    lam_vol = np.zeros((samples.shape[0],)) 
+    lam_vol = np.zeros((samples.shape[0],1)) 
     for i in range(samples.shape[0]):
         lam_vol[i] = np.sum(np.equal(emulate_ptr, i))
     clam_vol = np.copy(lam_vol) 
@@ -274,6 +283,7 @@ def prob_mc(samples, data, rho_D_M, d_distr_samples,
     lam_vol = lam_vol/(len(lambda_emulate)*size)
 
     # Calculate Probabilities
+    # TODO Fix issue of unecessary mpi call
     P = np.zeros((samples.shape[0],))
     for i in range(rho_D_M.shape[0]):
         Itemp = np.equal(io_ptr, i)
